Move default commands to Commands folder
